feat:rate, loss and teamimg
