+++ frontend/src/plugins/vuetify.js
@@ -20,7 +20,7 @@ export default createVuetify({
           primary: '#1867C0',
           secondary: '#5CBBF6',
           gamebgcolor: '#616161',
-          buttonbgcolor: '##212121'
+          buttonbgcolor: '#212121'
         },
       },
     },
